+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 #### Ymui
 * New widget: `YmuiMultiRippleView`.
 
+#### YmLib
+* New delegate class `SharedPreferenceDelegate`
+
 ### 0.6.0
 #### New annotation and processor support!
 * Generated application with application lifecycle observers